# Analyze frame-by-frame timing of stimuli