--- a/Analysis/debugFrameTime.m
+++ b/Analysis/debugFrameTime.m
@@ -1,9 +1,9 @@
-% It looks like the videos are shooting for a frame rate of 60 fps
-% So the expected duration is 0.0167 sec or 16.7 msec
-% But the measured duration of each 'frame' is closer to 0.669 msec,
-% Which means PTB is trying to refresh at a rate closer to 1000 fps
-% So this analysis is a bust.
-% I have no idea whether timing is significantly degraded or not.
+% Reads in ANY .tsv file in beh/ with 'task-debug' in the filename
+% Expects to see a column called 'Onset' for each frame of each video
+% Calculates frame duration based on onset times, 
+% then generates a histogram 
+
+% 
 
 pths = specifyPaths('..');
 % pths.beh = pwd;
@@ -17,33 +17,33 @@ for i = 1:numFiles
     numStims = length(stimList);
     output(i).name = fname;
     output(i).numStims = numStims;
-    figure();
+    flatOutput = [];
     for o = 1:numStims
-        output(i).stim(o).name = stimList(i);
+        output(i).stim(o).name = stimList(o);
         % We're dealing with onset times, but we care about durations
         % So use diff to get the differences b/w onset times, 
         % which is essentially the same thing as duration
         % Produces n-1 results since we don't get an offset for the final
-        output(i).stim(o).data = diff(input.Onset(strcmp(input.StimName,stimList(i))));
-
+        output(i).stim(o).data = diff(input.Onset(strcmp(input.StimName,stimList(o))));
+        flatOutput = [flatOutput; output(i).stim(o).data];
         % Subplot histograms for every movie
-        subplot(1,numStims,o);
+        figure();
             histogram(output(i).stim(o).data * 1000);
-            title(stimList(i));
+            title(strrep(stimList(o),'_','\_'));
             xlabel('Frame duration (msec)');
             ylabel('Num instances');
-            xlim([.5,1]); % fine-tune
+            xlim([0,36]); % fine-tune
             % ylim([0,100]); % fine-tune
     end
     % Separate by movie
     % data = pivot(input, Columns='StimName', Method="mean", DataVariable='Duration')
     % Plot histogram of frame durations
     figure()
-        histogram(input.Duration * 1000);
+        histogram(flatOutput * 1000);
         title(sprintf('Considering %i movies', numStims));
         xlabel('Frame duration (msec)');
         ylabel('Num instances');
-        xlim([.5,1]); % fine-tune
+        xlim([0,36]); % fine-tune
     
     % t-test for differences by movie?
 end